refactor: rename the TextInput component to InputField
