+++ resources/js/Pages/Auth/ConfirmPassword.vue
@@ -3,7 +3,7 @@ import GuestLayout from '@/Layouts/GuestLayout.vue';
 import InputError from '@/Components/InputError.vue';
 import InputLabel from '@/Components/InputLabel.vue';
 import PrimaryButton from '@/Components/PrimaryButton.vue';
-import TextInput from '@/Components/TextInput.vue';
+import InputField from '@/Components/InputField.vue';
 import { Head, useForm } from '@inertiajs/vue3';
 
 const form = useForm({
@@ -31,7 +31,7 @@ const submit = () => {
         <form @submit.prevent="submit">
             <div>
                 <InputLabel for="password" value="Password" />
-                <TextInput
+                <InputField
                     id="password"
                     type="password"
                     class="mt-1 block w-full"
