+++ resources/js/Pages/Auth/Register.vue
@@ -3,7 +3,7 @@ import GuestLayout from "@/Layouts/GuestLayout.vue";
 import InputError from "@/Components/InputError.vue";
 import InputLabel from "@/Components/InputLabel.vue";
 import PrimaryButton from "@/Components/PrimaryButton.vue";
-import TextInput from "@/Components/TextInput.vue";
+import InputField from "@/Components/InputField.vue";
 import { Head, Link, useForm } from "@inertiajs/vue3";
 
 const form = useForm({
@@ -40,7 +40,7 @@ const submit = () => {
             <div>
                 <InputLabel for="name" value="Name" />
 
-                <TextInput
+                <InputField
                     id="name"
                     type="text"
                     class="mt-1 block w-full"
@@ -56,7 +56,7 @@ const submit = () => {
             <div class="mt-4">
                 <InputLabel for="email" value="Email" />
 
-                <TextInput
+                <InputField
                     id="email"
                     type="email"
                     class="mt-1 block w-full"
@@ -71,7 +71,7 @@ const submit = () => {
             <div class="mt-4">
                 <InputLabel for="password" value="Password" />
 
-                <TextInput
+                <InputField
                     id="password"
                     type="password"
                     class="mt-1 block w-full"
@@ -89,7 +89,7 @@ const submit = () => {
                     value="Confirm Password"
                 />
 
-                <TextInput
+                <InputField
                     id="password_confirmation"
                     type="password"
                     class="mt-1 block w-full"
