+++ resources/js/Pages/Profile/Partials/UpdateProfileInformationForm.vue
@@ -2,7 +2,7 @@
 import InputError from '@/Components/InputError.vue';
 import InputLabel from '@/Components/InputLabel.vue';
 import PrimaryButton from '@/Components/PrimaryButton.vue';
-import TextInput from '@/Components/TextInput.vue';
+import InputField from '@/Components/InputField.vue';
 import { Link, useForm, usePage } from '@inertiajs/vue3';
 
 defineProps<{
@@ -37,7 +37,7 @@ const form = useForm({
             <div>
                 <InputLabel for="name" value="Name" />
 
-                <TextInput
+                <InputField
                     id="name"
                     type="text"
                     class="mt-1 block w-full"
@@ -53,7 +53,7 @@ const form = useForm({
             <div>
                 <InputLabel for="email" value="Email" />
 
-                <TextInput
+                <InputField
                     id="email"
                     type="email"
                     class="mt-1 block w-full"
